Minor changes in benchmarking code and experiments with JAX

+++ scripts/benchmark_rdm_environment.py
@@ -10,6 +10,7 @@ PATH = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.split(PATH)[0])
 from src.envs.rdm_environment import QubitsEnvironment
 
+BENCHMARK_NAME = 'benchmark-rdm-cython2'
 NSTEPS = 100
 REPEATS = 1
 results = {}
@@ -17,33 +18,28 @@ Ls = (4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14)
 Bs = (32, 64, 128, 256)
 
 os.makedirs('data/benchmarks/', exist_ok=True)
-# if os.path.exists('data/benchmarks/benchmark-rdm-environment-0.pkl'):
-#     with open('data/benchmarks/benchmark-rdm-environment-0.pkl', mode='rb') as f:
-#         results = pickle.load(f)
-# # else:
-for L in Ls:
-    for B in Bs:
-        print(f'Rollout with L={L}, B={B}...')
-        E = QubitsEnvironment(num_qubits=L, batch_size=B)
-        for _ in range(REPEATS):
-            E.set_random_state(copy=False)
-            for a in np.random.uniform(0, E.num_actions, (NSTEPS, B)).astype(np.int32):
-                state, rewards, done = E.step2(a)
-
-        for _ in range(REPEATS):
-            E.set_random_state(copy=False)
-            tick = time.time()
-            for a in np.random.uniform(0, E.num_actions, (NSTEPS, B)).astype(np.int32):
-                state, rewards, done = E.step2(a)
-            tock = time.time()
-            results.setdefault((L, B), []).append(tock - tick)
-print('Saving benchmark results...')
-with open('data/benchmarks/benchmark-rdm-environment-v5-[jax.eigh].pkl', mode='wb') as f:
-    pickle.dump(results, f)
+if os.path.exists(f'data/benchmarks/{BENCHMARK_NAME}.pkl'):
+    with open(f'data/benchmarks/{BENCHMARK_NAME}.pkl', mode='rb') as f:
+        results = pickle.load(f)
+else:
+    for L in Ls:
+        for B in Bs:
+            print(f'Rollout with L={L}, B={B}...')
+            E = QubitsEnvironment(num_qubits=L, batch_size=B)
+            for _ in range(REPEATS):
+                E.set_random_state(copy=False)
+                tick = time.time()
+                for a in np.random.uniform(0, E.num_actions, (NSTEPS, B)).astype(np.int32):
+                    state, rewards, done = E.step2(a)
+                tock = time.time()
+                results.setdefault((L, B), []).append(tock - tick)
+    print('Saving benchmark results...')
+    with open(f'data/benchmarks/{BENCHMARK_NAME}.pkl', mode='wb') as f:
+        pickle.dump(results, f)
 
 print('Creating figure...')
 fig, ax = plt.subplots(figsize=(12, 6))
-ax.set_title('v4 Optimized RDM Environment')
+ax.set_title(f'{BENCHMARK_NAME}')
 ax.set_xlabel('batch size')
 ax.set_ylabel('seconds')
 markers = '.o^+sxDv'
@@ -65,4 +61,4 @@ ax.set_xticks(Bs)
 ax.set_ylim(0, 5)
 ax.grid(True)
 ax.legend(loc='upper left')
-fig.savefig('data/benchmarks/benchmark-rdm-environment-v5-[jax.eigh].pdf')
+fig.savefig(f'data/benchmarks/{BENCHMARK_NAME}.pdf')
